% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Choup.R
\docType{data}
\name{Choup}
\alias{Choup}
\title{Coefficients de houppier.}
\format{
data frame contenant près de 425 lignes et 3 variables
}
\usage{
Choup
}
\description{
Coefficients de houppier issus de la base de données de l'AFI. Ils permettent de calculer le couvert
à partir de la surface terrière par essence et classe de diamètre.
}
\examples{
data(Choup)


}
\keyword{datasets}
